added config file feature and moved main method to drivers

+++ src/test/java/MapReduceTests.java
@@ -1,4 +1,3 @@
-import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.LongWritable;
 import org.apache.hadoop.io.Text;
 
@@ -21,49 +20,49 @@ public class MapReduceTests {
         String sampleInput = "Salam Salam Chetori? Khoobi?" +
                 " Salam Salamati Miare. Migi Chetori? Nemidoonam!";
 
-        new MapDriver<LongWritable, Text, Text, IntWritable>(new WordCountMapper())
+        new MapDriver<LongWritable, Text, Text, LongWritable>(new WordCountMapper())
                 .withInput(new LongWritable(0) ,new Text(sampleInput))
-                .withOutput(new Text("Salam"), new IntWritable(1))
-                .withOutput(new Text("Salam"), new IntWritable(1))
-                .withOutput(new Text("Chetori?"), new IntWritable(1))
-                .withOutput(new Text("Khoobi?"), new IntWritable(1))
-                .withOutput(new Text("Salam"), new IntWritable(1))
-                .withOutput(new Text("Salamati"), new IntWritable(1))
-                .withOutput(new Text("Miare."), new IntWritable(1))
-                .withOutput(new Text("Migi"), new IntWritable(1))
-                .withOutput(new Text("Chetori?"), new IntWritable(1))
-                .withOutput(new Text("Nemidoonam!"), new IntWritable(1))
+                .withOutput(new Text("Salam"), new LongWritable(1))
+                .withOutput(new Text("Salam"), new LongWritable(1))
+                .withOutput(new Text("Chetori?"), new LongWritable(1))
+                .withOutput(new Text("Khoobi?"), new LongWritable(1))
+                .withOutput(new Text("Salam"), new LongWritable(1))
+                .withOutput(new Text("Salamati"), new LongWritable(1))
+                .withOutput(new Text("Miare."), new LongWritable(1))
+                .withOutput(new Text("Migi"), new LongWritable(1))
+                .withOutput(new Text("Chetori?"), new LongWritable(1))
+                .withOutput(new Text("Nemidoonam!"), new LongWritable(1))
                 .runTest();
     }
 
     @Test
     public void testWordCountReducer() throws IOException {
-        new ReduceDriver<Text, IntWritable,Text, IntWritable>(new WordCountReducer())
-                .withInput(new Text("test") , Arrays.asList(new IntWritable(1),new IntWritable(1),new IntWritable(1)))
-                .withOutput(new Text("test"), new IntWritable(3))
+        new ReduceDriver<Text, LongWritable,Text, LongWritable>(new WordCountReducer())
+                .withInput(new Text("test") , Arrays.asList(new LongWritable(1),new LongWritable(1),new LongWritable(1)))
+                .withOutput(new Text("test"), new LongWritable(3))
                 .runTest();
     }
 
     @Test
     public void testReverseSortMap() throws IOException {
-        new MapDriver<Text, Text, IntWritable, Text>(new ReverseSortMapper())
+        new MapDriver<Text, Text, LongWritable, Text>(new ReverseSortMapper())
                 .withInput(new Text("Salam") , new Text("3"))
                 .withInput(new Text("Hello") , new Text("   2"))
-                .withOutput(new IntWritable(3), new Text("Salam"))
-                .withOutput(new IntWritable(2), new Text("Hello"))
+                .withOutput(new LongWritable(3), new Text("Salam"))
+                .withOutput(new LongWritable(2), new Text("Hello"))
                 .runTest();
 
     }
 
     @Test
     public void testReverseSortReducer() throws IOException {
-        new ReduceDriver<IntWritable, Text, Text, IntWritable>(new ReverseSortReducer())
-                .withInput(new IntWritable(2), Arrays.asList(new Text("Salam"),new Text("Hello")))
-                .withInput(new IntWritable(1), Arrays.asList(new Text("Marhaba"), new Text("Bonjour")))
-                .withOutput(new Text("Salam"), new IntWritable(2))
-                .withOutput(new Text("Hello"), new IntWritable(2))
-                .withOutput(new Text("Marhaba"), new IntWritable(1))
-                .withOutput(new Text("Bonjour"), new IntWritable(1))
+        new ReduceDriver<LongWritable, Text, Text, LongWritable>(new ReverseSortReducer())
+                .withInput(new LongWritable(2), Arrays.asList(new Text("Salam"),new Text("Hello")))
+                .withInput(new LongWritable(1), Arrays.asList(new Text("Marhaba"), new Text("Bonjour")))
+                .withOutput(new Text("Salam"), new LongWritable(2))
+                .withOutput(new Text("Hello"), new LongWritable(2))
+                .withOutput(new Text("Marhaba"), new LongWritable(1))
+                .withOutput(new Text("Bonjour"), new LongWritable(1))
                 .runTest();
     }
 }
